requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bertopic==0.15.0
bitarray==2.8.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Cython==0.29.36
deprecation==2.1.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fairseq==0.12.2
fastapi==0.104.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gotrue==1.3.0
gunicorn==21.2.0
h11==0.14.0
hdbscan==0.8.33
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.17.3
hydra-core==1.0.7
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwipiepy==0.16.1
kiwipiepy-model==0.16.0
lightning-utilities==0.9.0
llvmlite==0.41.1
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.26.1
omegaconf==2.0.6
orjson==3.9.10
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
plotly==5.18.0
portalocker==2.8.2
postgrest==0.13.0
protobuf==4.25.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pynndescent==0.5.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-lightning==1.8.5.post0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
realtime==1.0.0
regex==2023.10.3
requests==2.31.0
sacrebleu==2.3.2
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sklearn==0.0.post11
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.3
supafunc==0.3.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.0.0
torchaudio==2.0.1
torchmetrics==1.2.0
torchvision==0.15.1
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
umap-learn==0.5.4
urllib3==2.0.7
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==10.4
yarl==1.9.2
fcntl